Separate screeninfo.py into multiple files
This commit restructures the project significantly. Docstrings are added, copyright notices, functions moved to separate files, utils.py now contains load_library to prevent code repetition
